Adds default compute type to runtime env model, uses it (#197)
New option is called `default_compute_type`. If present, and if the list of `compute_types` contains it, we will display the default compute type when a particular environment is selected. If the default compute type is not specified or is not in the compute types list, we will display the first compute type in the list by default (as was the case before).