Add `_update_last_activity` to the extension app
Added `_update_last_activity` to the extension app to update the
`last_activity_times` look up in the web application. Pass this method
down to the scheduler so that it can be used to update the
`last_acitivity` whenever a job is created. May want to add it elsewhere
in the future, but this is a good start.